Working prototype
The ./add-on directory is the web extension side, which is a simple content
script which passes document html and metadata to the background script, which
passes it to the native application.

The native application is a python script in ./app which saves the content in
the required format to `~/.recollweb/ToIndex` (hardcoded for now).

This should reproduce the current behavior of the recoll-we extension.